Backend/ai_healthcare/requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
keras>=2.12.0
flask>=2.0.0
kaggle>=1.5.12
python-dotenv>=0.19.0
gunicorn>=20.1.0
flask-cors==3.0.10 